fix indexing issue in pairwise.complete, symmetrize

+++ src/utils/fill.h
@@ -46,6 +46,22 @@ static inline void diag2one(const int n, double *restrict x)
 
 
 // Copy lower triangle to upper
+// static inline void symmetrize(const int n, double *restrict x)
+// {
+//   const int blocksize = 8; // TODO check cache line explicitly
+  
+//   for (int j=0; j<n; j+=blocksize)
+//   {
+//     for (int i=j+1; i<n; i+=blocksize)
+//     {
+//       for (int col=j; col<j+blocksize && col<n; ++col)
+//       {
+//         for (int row=i; row<i+blocksize && row<n; ++row)
+//           x[col + n*row] = x[row + n*col];
+//       }
+//     }
+//   }
+// }
 static inline void symmetrize(const int n, double *restrict x)
 {
   const int blocksize = 8; // TODO check cache line explicitly
@@ -57,7 +73,7 @@ static inline void symmetrize(const int n, double *restrict x)
       for (int col=j; col<j+blocksize && col<n; ++col)
       {
         for (int row=i; row<i+blocksize && row<n; ++row)
-          x[col + n*row] = x[row + n*col];
+          x[col + (size_t)n*row] = x[row + (size_t)n*col];
       }
     }
   }
